+++ packages/core-api/lib/versions/2/methods/blocks.js
@@ -51,10 +51,14 @@ const search = async request => {
 }
 
 module.exports = server => {
+  const {
+    generateTimeout,
+  } = require('@arkecosystem/core-container').resolveOptions('api').cache
+
   server.method('v2.blocks.index', index, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -67,7 +71,7 @@ module.exports = server => {
   server.method('v2.blocks.show', show, {
     cache: {
       expiresIn: 600 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request => generateCacheKey({ id: request.params.id }),
@@ -76,7 +80,7 @@ module.exports = server => {
   server.method('v2.blocks.transactions', transactions, {
     cache: {
       expiresIn: 600 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -89,7 +93,7 @@ module.exports = server => {
   server.method('v2.blocks.search', search, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>