+++ packages/core-api/lib/versions/2/methods/transactions.js
@@ -35,10 +35,14 @@ const search = async request => {
 }
 
 module.exports = server => {
+  const {
+    generateTimeout,
+  } = require('@arkecosystem/core-container').resolveOptions('api').cache
+
   server.method('v2.transactions.index', index, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -51,7 +55,7 @@ module.exports = server => {
   server.method('v2.transactions.show', show, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request => generateCacheKey({ id: request.params.id }),
@@ -60,7 +64,7 @@ module.exports = server => {
   server.method('v2.transactions.search', search, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 8000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>